test: HandlerAdapter 테스트 코드 보충
- 구현체 별로 테스트 클래스 분리
- 부족한 테스트 케이스 보충
